+++ replication/src/main/java/com/opdar/gulosity/replication/base/StoreCallback.java
@@ -4,5 +4,5 @@ package com.opdar.gulosity.replication.base;
  * Created by 俊帆 on 2016/11/2.
  */
 public interface StoreCallback {
-    public void store(int position,int nextPosition);
+    public void store(long position,long nextPosition);
 }
