+++ replication/src/main/java/com/opdar/gulosity/replication/server/TcpServer.java
@@ -38,7 +38,7 @@ public class TcpServer implements StoreCallback{
     }
 
     @Override
-    public void store(int position, int nextPosition) {
+    public void store(long position, long nextPosition) {
         Registry.notifyClients(position,nextPosition);
     }
 }
